remote_bricks/requirements.txt
fastapi==0.111.0
passlib[bcrypt]==1.7.4
pydantic==2.7.4
pymongo[srv]==4.7.3
python-dotenv==1.0.1
uvicorn==0.30.1
